refactor: addressing pr comments
